Restructure notification service extension target & shared code
The Shared symlink was causing some issues and had duplicated files in a few places.

The OptimoveNotificationServiceExtension target still had Optipush code in it as well as the Kumulos SDK extension code.

Remove and add a new target for the extension that only contains the Kumulos SDK sources.

Renaming to follow.

Note this likely breaks packaging for pods etc., expect to pick up packaging in a follow-up PR.