Refactor SDK config to support both Optimove and Optimobile configurations
Promote and extend KSConfigBuilder as OptimoveConfigBuilder. It now expects to be given credentials in the base64-encoded JSON tuple provided in the mobile UI.

Refactor Kumulos SDK classes to expect an OptimobileConfig, held in the OptimoveConfig.

Add a new initialize method that conditionally inits the configured SDK behaviours.

Note there is follow-up work to merge the public API of Optimobile into the Optimove class, so it is commented out in this build.