Use Optimove initialVisitorId for Optimobile installId
Refactor the Kumulos installId to use the Optimove initialVisitorId instead.

The initialVisitorId is also a UUIDv4 so this change should be transparent to the Optimobile implementation.

In order to achieve this in cases where Optimove credentials are not yet provided, we need to re-arrange the initialisation of some dependencies in the container.

Move the Logger init from configure, and the new visitor ID generation from startSdk into the Optimove init() method instead.