Don't block caller when performing in-app syncs
Only block the serial in-app sync queue's thread.

This API already supports asynchronous completion notification via the block so it isn't necessary
to block the caller. All callsites that are interested in the result use the completion block already.

This still ensures there can only be one sync operation being performed at a time, and all state will
be consistent if multiple syncs are triggered concurrently.

Whilst the sync task block will wait indefinitely for the networking and persistence operations to
complete, in practice this should be unlikely. Failed or timing out network operations will trigger
onFailure and signal the semaphore, and storage operations are local and failures are silent.