Don't perform unnecessary in-app syncs from content-available pushes
In the previous implementation, any content-available push notification would trigger an in-app
sync operation, regardless of if the notification contained an in-app deep link payload.

This refactoring adds conditional checks to ensure that in-app syncs are only performed when they
are necessary.

In addition, the synchronisation logic between in-app syncs and any user's background notification
application delegate is improved to reduce the chance of timeouts on the main thread from the semaphore.

Now we use a DispatchGroup to fork and join the two units of work (user delegate + in-app sync) with no
lengthy blocking of the main thread.

To ensure the same semantics for the user, the invocation of the existing app delegate is still performed
on the main thread, but the work is queued to allow setting up of the DispatchGroup.

Note this refactoring has removed support for iOS 9 push notifications in this handler, but this is likely
acceptable as the minimum target for the Optimove SDK is iOS 10, and we are now on iOS 15.