Reduce wait time on events semaphore
When the app is terminating, we try to flush the events queue and we wait on the main thread to
prevent the process exiting. However, under certain conditions it seems the watchdog process
will uncleanly kill the app because the app didn't terminate within 5 seconds.

As a solution, reduce the maximum time we will wait for syncing to happen to 3 seconds.