Remove redundant in-app syncs during initialisation
Split the storage of last sync time from most recent message updated at.

The sync API needs to know the most recent update time of the client, but that could
be distinct from the client's last sync timestamp (i.e. when the network call was made).

By using the most recent message update time to try and debounce syncing in the foreground
handler, it led to redundant syncs on app startup. The initial sync from enrollment setup
would happen, but then if a client has never received a message, the most recent update time
was nil, so the foreground handler would immediately sync again.

Despite there being a guard that looked like it would address this, the guard was ineffective
due to queueing work async on the main thread, by which time the app state was active. The
solution of splitting the client's last sync time from the sync's most recent update helps
remove the need for this guard.

Remove registered guard from enrollment and sync setup.

This flag is not present in the reference Objective-C implementation, and its presence in
the Swift imlpementation introduces a subtle bug where manual consent management is used.

In this case, opting in for in-app relies on handleEnrollmentAndSyncSetup to re-fetch the
user's messages when they opt in. Because of the 'registered' flag, a user who opted out
and in again in the same session (perhaps unlikely) would not fetch until their next cold
start.