Enable in-app auto-enrollment when upgrading installs from Kumulos to Optimove SDKs
The Optimobile system uses the same storage key for in-app consent as the Kumulos SDK did previously. This means in-app auto-enrollment doesn't happen when an install updates from the Kumulos to the Optimove SDK because the SDK thinks the user is already consented.

To resolve this, we detect if we are a 'new' Optimobile installation and remove the in-app consent key from storage. This enables the typical in-app consent process to take place for auto-enrollment.

Note that explicit consent management use-cases are not considered and they would end up in a non-consented state until the opt-in is sent again by the app.